Add an option to omit BOS/EOS tokens from outputs
